// src/context/enum.ts
export enum Action {
  OPEN_SNACKBAR = 'OPEN_SNACKBAR',
  CLOSE_SNACKBAR = 'CLOSE_SNACKBAR',
  CHANGE_WALLET_DATE = 'CHANGE_WALLET_DATE',
  OPEN_CATEGORY_DIALOG = 'CATEGORY_DIALOG',
  CLOSE_CATEGORY_DIALOG = 'CLOSE_CATEGORY_DIALOG',
  DIALOG_DELETION_OPEN = 'DIALOG_DELETION_OPEN',
  DIALOG_DELETION_CLOSE = 'DIALOG_DELETION_CLOSE',
  IS_DELETING_BUDGET = 'IS_DELETING_BUDGET',
  BUDGET_HAS_BEEN_DELETED = 'BUDGET_HAS_BEEN_DELETED',
  CLEAR_GLOBAL_STATE = 'CLEAR_GLOBAL_STATE',
  DIALOG_EDITION_OPEN = 'DIALOG_EDITION_OPEN',
  DIALOG_EDITION_CLOSE = 'DIALOG_EDITION_CLOSE'
}